Refactor model and service interfaces for bridges and schedules
- Introduced a new `BridgeProvider` interface to encapsulate bridge-related operations.
- Updated the `Service` struct to use the new `BridgeProvider` interface.
- Implemented methods for loading, saving, and removing bridges in the `Provider` struct.
- Refactored the `ScheduleProvider` interface to include methods for schedule management.
- Updated the `Service` struct to utilize the new `ScheduleProvider` interface.
- Refactored the HTTP server to interact with the new service structure.
- Updated tests to reflect changes in the service and provider interfaces.
